feat: implement product authorization policies and update ProductController for access control

+++ app/Http/Controllers/ProductController.php
@@ -6,12 +6,14 @@ use App\Http\Requests\ProductsRequest;
 use App\Models\Brand;
 use App\Models\Category;
 use App\Models\Product;
+use App\Models\ProductTags;
 use App\Models\Tags;
 use App\Models\User;
 use Illuminate\Container\Attributes\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
 class ProductController extends Controller
@@ -66,6 +68,7 @@ class ProductController extends Controller
 
     public function edit(Product $product)
     {
+        Gate::authorize('update', $product);
         $categories = Category::all();
         $brands = Brand::all();
         return view('Products.edit', compact('product', 'categories', 'brands'));
@@ -73,6 +76,8 @@ class ProductController extends Controller
 
     public function update(Request $request, Product $product)
     {
+        Gate::authorize('update', $product);
+
         $product->brand_id = $request->brand_id;
         $product->category_id = $request->category_id;
         $product->name = $request->name;
@@ -107,6 +112,10 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
+
+        Gate::authorize('delete', $product);
+
+        ProductTags::where('product_id', $product->id)->delete();
         $product->delete();
         return redirect()->route('products.index')->with('success', 'Product deleted successfully.');
     }
